--- react-frontend/src/components/newFile.jsx
@@ -1,3 +1,0 @@
-import { Navbar } from "./Navbar";
-
-export default Navbar;
